feat: enhanced -R / --reference option
Previously, in order to get reference information on a specific
directive (let's say `!id`) one would run

    infodisplay -R !id

The problem is that `!` is already used in bash so bash reports an
error. I didn't see this problem because I use fish as my daily driver.

In order to command to work, it would have to be typed out like this:

    infodisplay -R \!id

Which is ugly and annoying. Instead, I have added the option to omit the
leading `!`. So both `-Rid` and `-R!id` work now.

I have also added an error message if there is no reference

Signed-off-by: Ben Raz <ben.raz2008@gmail.com>